rowIndex is set incorrectly on the last page.
rowIndex is set incorrectly on the last page, as the last page might have fewer rows than pagesize. For example, if a table has 33 table rows and we are on page 4.
Then pageRows.size would be 3 and rowindex of the first row on page 4 would be 10 but the correct value is 31.